chore: remove KernelMethod base class to add it back later
