+++ qlkit/algorithms/__init__.py
@@ -1,12 +1,10 @@
 from .quantum_estimator import QuantumEstimator
-from .kernel_method import KernelMethod
 from .qknn import QKNeighborsClassifier
 from .qsvm import QSVClassifier
 from .qkmeans import QKMeans
 
 __all__ = [
     "QuantumEstimator",
-    "KernelMethod",
     "QKNeighborsClassifier",
     "QSVClassifier",
     "QKMeans"
